+++ packages/build-tools/src/steps/utils/ios/configure.ts
@@ -6,6 +6,7 @@ import uniq from 'lodash/uniq';
 import path from 'path';
 
 import { Credentials } from './credentials/manager';
+import { DistributionType, ProvisioningProfileData } from './credentials/provisioningProfile';
 
 export async function configureCredentialsAsync(
   logger: bunyan,
@@ -31,6 +32,30 @@ export async function configureCredentialsAsync(
       buildConfiguration,
     });
   }
+
+  if (credentials.distributionType === DistributionType.DEVELOPMENT) {
+    overrideCodeSignIdentityForDevelopment(
+      workingDir,
+      credentials.targetProvisioningProfiles,
+      buildConfiguration
+    );
+  }
+}
+
+function overrideCodeSignIdentityForDevelopment(
+  projectRoot: string,
+  targetProvisioningProfiles: Record<string, ProvisioningProfileData>,
+  buildConfiguration: string
+): void {
+  const project = IOSConfig.XcodeUtils.getPbxproj(projectRoot);
+  for (const [targetName, profile] of Object.entries(targetProvisioningProfiles)) {
+    const xcBuildConfiguration = IOSConfig.Target.getXCBuildConfigurationFromPbxproj(project, {
+      targetName,
+      buildConfiguration,
+    });
+    xcBuildConfiguration.buildSettings.CODE_SIGN_IDENTITY = `"${profile.certificateCommonName}"`;
+  }
+  fs.writeFileSync(project.filepath, project.writeSync());
 }
 
 export async function updateVersionsAsync(